requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
Faker==27.4.0
Markdown==3.7
pillow==10.4.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
